requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pandas==2.2.2
numpy==1.26.4
opencv-python==4.9.0.80
Pillow==10.3.0
ultralytics==8.3.133 
scikit-learn==1.4.2
matplotlib==3.8.4
